authentication and recipe editor by only owner

+++ app/Http/Controllers/RecipeController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Recipe;
+use Auth;
+use App\User;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\PostFormRequest;
 
 class RecipeController extends Controller
 {
@@ -33,21 +36,20 @@ class RecipeController extends Controller
     public function store(Request $request)
     {
         //
-
     $this->validate($request, [
         'name' => 'required',
         'description' => 'required',
         'difficult' => 'required'
     ]);
-
-    $input = $request->all();
-
-     Recipe::create($input);
-
+    //$input = $request->all();
+     //Recipe::create($input);
+     $input = new Recipe($request->all());
+    Auth::user()->recipe()->save($input);
     Session()->flash('flash_message', 'Ricetta aggiunta con successo!');
-
     return redirect()->back();
     }
+
+    
 
     /**
      * Show the form for editing the specified resource.
@@ -55,11 +57,16 @@ class RecipeController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Request $request, $id)
     {
+        
         $recipe = Recipe::findOrFail($id);
+        if($recipe && ($request->user()->id == $recipe->user_id || $request->user()->is_admin())){
+        return view('recipe.edit')->withRecipe($recipe);
+        }
+        return redirect('/recipe')->withErrors('you have not sufficient permissions');
 
-    return view('recipe.edit')->withRecipe($recipe);
+    //return view('recipe.edit')->withRecipe($recipe);
     }
 
     /**
@@ -73,19 +80,24 @@ class RecipeController extends Controller
     {
         $recipe = Recipe::findOrFail($id);
 
-    $this->validate($request, [
+        $this->validate($request, [
         'name' => 'required',
         'description' => 'required',
         'difficult' => 'required'
     ]);
 
-    $input = $request->all();
-
-    $recipe->fill($input)->save();
-
-    Session()->flash('flash_message', 'Aggiornato correttamente');
-
-    return redirect()->back();
+        $input = $request->all();
+        if($recipe && ($request->user()->id == $recipe->user_id || $request->user()->is_admin()))
+    {
+        $recipe->fill($input)->save();
+    
+        Session()->flash('flash_message', 'Aggiornato correttamente');
+    
+        return redirect()->back();
+    }else
+        {
+          return redirect('/home')->withErrors('you have not sufficient permissions');
+        }
     }
 
     /**
@@ -97,11 +109,17 @@ class RecipeController extends Controller
     public function destroy($id)
     {
         $recipe = Recipe::findOrFail($id);
-
-    $recipe->delete();
-
-    Session()->flash('flash_message', 'Cancellato');
-
-    return redirect()->route('recipe.index');
+    if($recipe && ($request->user()->id == $recipe->user_id || $request->user()->is_admin()))
+    {
+        $recipe->delete();
+    
+        Session()->flash('flash_message', 'Cancellato');
+    
+        return redirect()->route('recipe.index');
+    }else 
+    {
+        Session()->flash('flash_message', 'Non hai i permessi');
     }
+    }
+    
 }
